+++ services/argus/model_create_scrape_config_payload_metrics_relabel_configs_inner.go
@@ -14,7 +14,7 @@ type CreateScrapeConfigPayloadMetricsRelabelConfigsInner struct {
 	// Action to perform based on regex matching. `Additional Validators:` * if action is replace, targetLabel needs to be in body
 	Action *string `json:"action,omitempty"`
 	// Modulus to take of the hash of the source label values.
-	Modulus *float32 `json:"modulus,omitempty"`
+	Modulus *float64 `json:"modulus,omitempty"`
 	// Regular expression against which the extracted value is matched.
 	Regex *string `json:"regex,omitempty"`
 	// Replacement value against which a regex replace is performed if the regular expression matches. Regex capture groups are available.
