+++ services/dns/model_update_zone_payload.go
@@ -16,19 +16,19 @@ type UpdateZonePayload struct {
 	// contact e-mail for the zone
 	ContactEmail *string `json:"contactEmail,omitempty"`
 	// default time to live
-	DefaultTTL *int32 `json:"defaultTTL,omitempty"`
+	DefaultTTL *int64 `json:"defaultTTL,omitempty"`
 	// description of the zone
 	Description *string `json:"description,omitempty"`
 	// expire time
-	ExpireTime *int32 `json:"expireTime,omitempty"`
+	ExpireTime *int64 `json:"expireTime,omitempty"`
 	// user given name
 	Name *string `json:"name,omitempty"`
 	// negative caching
-	NegativeCache *int32 `json:"negativeCache,omitempty"`
+	NegativeCache *int64 `json:"negativeCache,omitempty"`
 	// primary name server for secondary zone
 	Primaries *[]string `json:"primaries,omitempty"`
 	// refresh time
-	RefreshTime *int32 `json:"refreshTime,omitempty"`
+	RefreshTime *int64 `json:"refreshTime,omitempty"`
 	// retry time
-	RetryTime *int32 `json:"retryTime,omitempty"`
+	RetryTime *int64 `json:"retryTime,omitempty"`
 }
