+++ services/mongodbflex/model_partial_update_instance_payload.go
@@ -18,7 +18,7 @@ type PartialUpdateInstancePayload struct {
 	Labels   *map[string]string `json:"labels,omitempty"`
 	Name     *string            `json:"name,omitempty"`
 	Options  *map[string]string `json:"options,omitempty"`
-	Replicas *int32             `json:"replicas,omitempty"`
+	Replicas *int64             `json:"replicas,omitempty"`
 	Storage  *InstanceStorage   `json:"storage,omitempty"`
 	Version  *string            `json:"version,omitempty"`
 }
