+++ services/postgresflex/model_user_response_user.go
@@ -13,7 +13,7 @@ package postgresflex
 type UserResponseUser struct {
 	Host     *string   `json:"host,omitempty"`
 	Id       *string   `json:"id,omitempty"`
-	Port     *int32    `json:"port,omitempty"`
+	Port     *int64    `json:"port,omitempty"`
 	Roles    *[]string `json:"roles,omitempty"`
 	Username *string   `json:"username,omitempty"`
 }
